+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Added `filterByLanguagePresets` (`QueryResultService`)
+* Improved `filterBySysLanguage` (`QueryResultService`)
+
 1.7.0
 -----